RETRY ncc1701kb with all commits and updates from former PR
RETRY of PR #22705 After a strange problem with the develop branch. It was causing a HUGE amount of conflict errors. I was told to just make a new PR. This contains ALL of the suggestions and commits from the reviewers of the form PR, so there shouldn't be any changes.